requirements.txt
discord.py==1.7.2
psycopg2==2.8.6